+++ src/ua/data_types/create_monitored_items_request.rs
@@ -5,7 +5,7 @@ crate::data_type!(CreateMonitoredItemsRequest);
 impl CreateMonitoredItemsRequest {
     #[must_use]
     pub const fn with_subscription_id(mut self, subscription_id: ua::SubscriptionId) -> Self {
-        self.0.subscriptionId = subscription_id.as_u32();
+        self.0.subscriptionId = subscription_id.as_id().as_u32();
         self
     }