+++ src/ua/data_types/create_subscription_respones.rs
@@ -7,7 +7,7 @@ crate::data_type!(CreateSubscriptionResponse);
 impl CreateSubscriptionResponse {
     #[must_use]
     pub const fn subscription_id(&self) -> ua::SubscriptionId {
-        ua::SubscriptionId::new(self.0.subscriptionId)
+        ua::SubscriptionId::new(ua::IntegerId::new(self.0.subscriptionId))
     }
 
     /// Gets revised publishing interval.